Use a private temp directory for certificate
Although this is likely to be on a somewhat restricted host, I still
don't like the thought of having a temporary file with the private key
somewhere that would be viewable by a normal user.  Create a temporary
directory to store the generated files which will only be accessible to
the user running the script.  This also cleans up after itself.