Partial fix for haproxy config
This now at least seems to be working correctly for virtual domains.
Previously a request for http://ncsysadmin.org would get forwarded to
port 443 and show up in /var/log/steering.trilug.org on the VM.

Still need to get certificates setup in haproxy so that https
connections can be handled correctly.